Checker program was added and tested
The checker program (bonus) appears to be working. It mainly
consists of the files: ch_main.c and ps_lang_interpreter.c. This
program shares source code with the push_swap program. It is yet
to be decided and enforced the bonus policy with regard to the
file naming (_bonus) and the rule to compile shared objects. The
documentation was improved and updated.