IMPROVEMENT Now the listed entity is a t_element
The previous dlclst_int_utils collection with a simple integer as the
list content type, was replaced by the new dlclst_element_utils
collection which defines a more complex t_element struct type and the
functions to manipulate it. The de-referenced list member content
that resulted in the actual number value, now is storaged as the
value member of the t_element.
Previously, given a is the content member of a node, the following:
*(int *)a
Is now equivalent to:
((t_element *)a)->value