Fully working version
Costs calculations functions finished. Both mandatory and bonus
versions work perfectly. Several test performed on both binaries:
malformed arguments; wrong instructions for the checker; valgrind
on normal and abnormal conditions.

While for a few tests the subject benchmark requirements were
succesfully achieved, it is still pending a broader set of numbers
and a statistically significant conclusion. For this, I'm thinking
of writing a test program which is already in the pseudo-code stage.