Tester almost done
This may be the final tester version. Some improvements like
retrieving and parsing the output of checker for KO or ERROR
instead of printing it on the screen. In this line, the resulting
instructions file of a push_swap execution should also be parsed
for ERROR, currently it is assumed that no ERROR is possible.

Additionally, a couple of particular random generated number sets
appeared to have crashed push_swap. These sets were retrieved as
files WEIRD_SET_1 and WEIRD_SET_2 for further analysis.