FIX2 infinite ra() loop in pre_sort() watchdog
A watchdog function was added in pre_sort() to prevent any possible
condition where the number of pushable elements left in stack A are
less than that the number required to be pb()'ed in a block
processing. This used to cause an infinite loop of ra()'s in the
pre-sort stage for some random numbers combinations.